Fix UX_DISPLAY() usage
The second argument to UX_DISPLAY() is a callback function with a
prototype of:

  const bagl_element_t *callback(const bagl_element_t *element)

we were passing a function like:

  unsigned int callback(const bagl_element_t *element)

When we used to return 0, we should return NULL instead to denote "do
not display".

When we used to return 1 (a pointer is expected!), we should return the
callback's argument to denote "no need to change, display as is".