chore: remove double-hash method
This commit removes the `doubleHash256` method, and updates `base58.c` to reflect the change.

Double hashing is not going to be needed very much, especially after current and future refactoring.

It's used in deriving a wallet address from a publicKey or RecipientId bytes, and is
only called from the `encodeBase58PublicKey` method.

This change is non-breaking.