docs(vagrant): update install steps
VirtualBox 6.1.2 is currently incompatible with the latest Vagrant (2.2.6).

Homebrew also unexpectedly does not have a clean or quick way to force an older version of VirtualBox without modifying system files or the homebrew cask.

While I expect this to eventually be resolved, I have no idea when that will happen.

This PR recommends manual installation of version-specific packages to replace the homebrew and apt install steps for macOS and Linux.

* VirtualBox 6.0.14
* VirtualBox Extensions Pack 6.0.14
* Vagrant 2.2.6